+++ modules/components/Query.jsx
@@ -1,26 +1,22 @@
-import React, { PureComponent } from "react";
+import React, { Component } from "react";
+import {connect} from "react-redux";
+import context from "../stores/context";
 import PropTypes from "prop-types";
 import * as actions from "../actions";
-import {fixPathsInTree} from "../utils/treeUtils";
 import {immutableEqual} from "../utils/stuff";
-import {useOnPropsChanged, bindActionCreators} from "../utils/reactUtils";
-import {validateTree} from "../utils/validation";
+import {useOnPropsChanged, liteShouldComponentUpdate, bindActionCreators} from "../utils/reactUtils";
 
 
-export const validateAndFixTree = (newTree, _oldTree, newConfig, oldConfig) => {
-  let tree = validateTree(newTree, _oldTree, newConfig, oldConfig);
-  tree = fixPathsInTree(tree);
-  return tree;
-};
-
-
-export default class Query extends PureComponent {
+class Query extends Component {
   static propTypes = {
     config: PropTypes.object.isRequired,
     onChange: PropTypes.func,
     renderBuilder: PropTypes.func,
     tree: PropTypes.any, //instanceOf(Immutable.Map)
     //dispatch: PropTypes.func.isRequired,
+    //__isInternalValueChange
+    //__lastAction
+    //getMemoizedTree: PropTypes.func.isRequired,
   };
 
   constructor(props) {
@@ -29,12 +25,13 @@ export default class Query extends PureComponent {
 
     this._updateActions(props);
 
-    this.validatedTree = this.validateTree(props, props);
-    //props.onChange && props.onChange(this.validatedTree, props.config);
-  }
+    // For preventive validation (tree and config consistency)
+    // When config has chnaged from QueryContainer, 
+    //  but new dispatched validated tree value is not in redux store yet (tree prop is old)
+    this.validatedTree = props.getMemoizedTree(props.config, props.tree);
+    this.oldValidatedTree = this.validatedTree;
 
-  validateTree (props, oldProps) {
-    return validateAndFixTree(props.tree, oldProps.tree, props.config, oldProps.config);
+    //props.onChange && props.onChange(this.validatedTree, props.config);
   }
 
   _updateActions (props) {
@@ -42,20 +39,32 @@ export default class Query extends PureComponent {
     this.actions = bindActionCreators({...actions.tree, ...actions.group, ...actions.rule}, config, dispatch);
   }
 
+  shouldComponentUpdate = liteShouldComponentUpdate(this, {
+    tree: (nextValue) => {
+      if (nextValue === this.oldValidatedTree && this.oldValidatedTree === this.validatedTree) {
+        // Got value dispatched from QueryContainer
+        // Ignore, because we've just rendered it
+        return false;
+      }
+      return true;
+    }
+  });
+
   onPropsChanged(nextProps) {
     const {onChange} = nextProps;
     const oldConfig = this.props.config;
     const newTree = nextProps.tree;
+    const oldTree = this.props.tree;
     const newConfig = nextProps.config;
-    const oldValidatedTree = this.validatedTree;
 
+    this.oldValidatedTree = this.validatedTree;
     this.validatedTree = newTree;
     if (oldConfig !== newConfig) {
       this._updateActions(nextProps);
-      this.validatedTree = this.validateTree(nextProps, this.props);
+      this.validatedTree = nextProps.getMemoizedTree(newConfig, newTree, oldConfig);
     }
 
-    const validatedTreeChanged = !immutableEqual(this.validatedTree, oldValidatedTree);
+    const validatedTreeChanged = !immutableEqual(this.validatedTree, this.oldValidatedTree);
     if (validatedTreeChanged) {
       onChange && onChange(this.validatedTree, newConfig, nextProps.__lastAction);
     }
@@ -74,3 +83,23 @@ export default class Query extends PureComponent {
     return renderBuilder(builderProps);
   }
 }
+
+
+const ConnectedQuery = connect(
+  (state) => {
+    return {
+      tree: state.tree,
+      __isInternalValueChange: state.__isInternalValueChange,
+      __lastAction: state.__lastAction,
+    };
+  },
+  null,
+  null,
+  {
+    context
+  }
+)(Query);
+ConnectedQuery.displayName = "ConnectedQuery";
+
+
+export default ConnectedQuery;