+++ modules/components/QueryContainer.jsx
@@ -1,36 +1,16 @@
 import React, { Component, PureComponent } from "react";
 import PropTypes from "prop-types";
-import createTreeStore from "../stores/tree";
+import treeStoreReducer from "../stores/tree";
 import context from "../stores/context";
 import {createStore} from "redux";
-import {connect, Provider} from "react-redux";
+import {Provider} from "react-redux";
 import * as actions from "../actions";
-import {extendConfig} from "../utils/configUtils";
-import {shallowEqual, immutableEqual} from "../utils/stuff";
+import {createConfigMemo} from "../utils/configUtils";
+import {immutableEqual} from "../utils/stuff";
 import {defaultRoot} from "../utils/defaultUtils";
+import {createValidationMemo} from "../utils/validation";
 import {liteShouldComponentUpdate, useOnPropsChanged} from "../utils/reactUtils";
-import pick from "lodash/pick";
-import Query, {validateAndFixTree} from "./Query";
-
-
-const configKeys = ["conjunctions", "fields", "types", "operators", "widgets", "settings", "funcs"];
-
-
-const ConnectedQuery = connect(
-  (state) => {
-    return {
-      tree: state.tree,
-      __isInternalValueChange: state.__isInternalValueChange,
-      __lastAction: state.__lastAction,
-    };
-  },
-  null,
-  null,
-  {
-    context
-  }
-)(Query);
-ConnectedQuery.displayName = "ConnectedQuery";
+import ConnectedQuery from "./Query";
 
 
 export default class QueryContainer extends Component {
@@ -52,16 +32,19 @@ export default class QueryContainer extends Component {
     super(props, context);
     useOnPropsChanged(this);
 
-    const config = pick(props, configKeys);
-    const extendedConfig = extendConfig(config);
+    this.getMemoizedConfig = createConfigMemo();
+    this.getMemoizedTree = createValidationMemo();
+    
+    const config = this.getMemoizedConfig(props);
     const tree = props.value;
-    const validatedTree = tree ? validateAndFixTree(tree, null, config, config) : null;
+    const validatedTree = this.getMemoizedTree(config, tree);
 
-    const store = createTreeStore({...config, tree: validatedTree});
+    const reducer = treeStoreReducer(config, validatedTree, this.getMemoizedTree);
+    const store = createStore(reducer);
 
     this.state = {
-      store: createStore(store),
-      config: extendedConfig,
+      store,
+      config
     };
   }
 
@@ -71,23 +54,24 @@ export default class QueryContainer extends Component {
 
   onPropsChanged(nextProps) {
     // compare configs
-    const oldConfig = pick(this.props, configKeys);
-    let nextConfig = pick(nextProps, configKeys);
-    const isConfigChanged = !shallowEqual(oldConfig, nextConfig, true);
-    if (isConfigChanged) {
-      nextConfig = extendConfig(nextConfig);
-      this.setState({config: nextConfig});
-    }
-    
+    const oldConfig = this.state.config;
+    const nextConfig = this.getMemoizedConfig(nextProps);
+    const isConfigChanged = oldConfig !== nextConfig;
+
     // compare trees
     const storeValue = this.state.store.getState().tree;
     const isTreeChanged = !immutableEqual(nextProps.value, this.props.value) && !immutableEqual(nextProps.value, storeValue);
-    if (isTreeChanged) {
-      const nextTree = nextProps.value || defaultRoot({ ...nextProps, tree: null });
-      const validatedTree = validateAndFixTree(nextTree, null, nextConfig, oldConfig);
+    const currentTree = isTreeChanged ? nextProps.value || defaultRoot(nextProps) : storeValue;
+
+    if (isConfigChanged) {
+      this.setState({config: nextConfig});
+    }
+    
+    if (isTreeChanged || isConfigChanged) {
+      const validatedTree = this.getMemoizedTree(nextConfig, currentTree, oldConfig);
       return Promise.resolve().then(() => {
         this.state.store.dispatch(
-          actions.tree.setTree(nextProps, validatedTree)
+          actions.tree.setTree(nextConfig, validatedTree)
         );
       });
     }
@@ -103,8 +87,8 @@ export default class QueryContainer extends Component {
       <QueryWrapper config={config}>
         <Provider store={store} context={context}>
           <ConnectedQuery
-            store={store}
             config={config}
+            getMemoizedTree={this.getMemoizedTree}
             onChange={onChange}
             renderBuilder={renderBuilder || get_children}
           />