+++ modules/components/widgets/mui/value/MuiTime.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import TimePicker from "@mui/lab/TimePicker";
+import { TimePicker } from "@mui/x-date-pickers/TimePicker";
 import moment from "moment";
 import FormControl from "@mui/material/FormControl";
 import TextField from "@mui/material/TextField";