+++ modules/config/default.js
@@ -48,7 +48,8 @@ export const settings = {
   forceShowConj: false,
   canShortMongoQuery: true,
   removeEmptyGroupsOnLoad: true,
-  removeIncompleteRulesOnLoad: false,
+  removeIncompleteRulesOnLoad: true,
+  removeInvalidMultiSelectValuesOnLoad: true,
   groupActionsPosition: "topRight", // oneOf [topLeft, topCenter, topRight, bottomLeft, bottomCenter, bottomRight]
   setOpOnChangeField: ["keep", "default"], // 'default' (default if present), 'keep' (keep prev from last field), 'first', 'none'
   groupOperators: ["some", "all", "none"],