+++ modules/export/elasticSearch.js
@@ -77,7 +77,7 @@ function buildEsRangeParameters(value, operator) {
 
   case "greater":
     return {
-      gte: "".concat(dateTime)
+      gt: "".concat(dateTime)
     };
 
   default:
@@ -103,22 +103,23 @@ function buildEsWildcardParameters(value) {
  * returns the ES occurrence required for bool queries
  *
  * @param {string} combinator - query group type or rule condition
+ * @param {bool} not
  * @returns {string} - ES occurrence type. See constants.js
  * @private
  */
-function determineOccurrence(combinator) {
+function determineOccurrence(combinator, not) {
   //todo: move into config, like mongoConj
   switch (combinator) {
   case "AND":
-    return "must";
+    return not ? "must_not" : "must";
     // -- AND
 
   case "OR":
-    return "should";
+    return not ? "should_not" : "should";
     // -- OR
 
   case "NOT":
-    return "must_not";
+    return not ? "must" : "must_not";
     // -- NOT AND
 
   default:
@@ -236,6 +237,8 @@ function buildEsRule(fieldName, value, operator, config, valueSrc) {
   // handle if value 0 has multiple values like a select in a array
   const widget = getWidgetForFieldOp(config, fieldName, op, valueSrc);
   const widgetConfig = config.widgets[widget];
+  if (!widgetConfig)
+    return undefined; // unknown widget
   const { elasticSearchFormatValue } = widgetConfig;
 
   /** In most cases the queryType will be static however in some casese (like between) the query type will change
@@ -281,15 +284,16 @@ function buildEsRule(fieldName, value, operator, config, valueSrc) {
  *
  * @param {object} children - The contents of the group
  * @param {string} conjunction - The way the contents of the group are joined together i.e. AND OR
+ * @param {bool} not
  * @param {Function} recursiveFxn - The recursive fxn to build the contents of the groups children
  * @private
  * @returns {object} - The ES group
  */
-function buildEsGroup(children, conjunction, recursiveFxn, config) {
+function buildEsGroup(children, conjunction, not, recursiveFxn, config) {
   if (!children || !children.size)
     return undefined;
   const childrenArray = children.valueSeq().toArray();
-  const occurrence = determineOccurrence(conjunction);
+  const occurrence = determineOccurrence(conjunction, not);
   const result = childrenArray.map((c) => recursiveFxn(c, config)).filter(v => v !== undefined);
   if (!result.length)
     return undefined;
@@ -331,10 +335,11 @@ export function elasticSearchFormat(tree, config) {
   }
 
   if (type === "group" || type === "rule_group") {
+    const not = properties.get("not");
     let conjunction = properties.get("conjunction");
     if (!conjunction)
       conjunction = defaultConjunction(config);
     const children = tree.get("children1");
-    return buildEsGroup(children, conjunction, elasticSearchFormat, config);
+    return buildEsGroup(children, conjunction, not, elasticSearchFormat, config);
   }
 }