+++ tests/specs/InteractionsVanilla.test.js
@@ -6,8 +6,8 @@ import { with_qb } from "../support/utils";
 
 
 describe("interactions on vanilla", () => {
-  it("click on remove single rule will leave empty rule if canLeaveEmptyGroup=false", () => {
-    with_qb(configs.dont_leave_empty_group, inits.with_number, "JsonLogic", (qb, onChange) => {
+  it("click on remove single rule will leave empty rule if canLeaveEmptyGroup=false", async () => {
+    await with_qb(configs.dont_leave_empty_group, inits.with_number, "JsonLogic", (qb, onChange) => {
       qb
         .find(".rule .rule--header button")
         .first()
@@ -22,8 +22,8 @@ describe("interactions on vanilla", () => {
     });
   });
 
-  it("click on remove group will leave empty rule if canLeaveEmptyGroup=false", () => {
-    with_qb(configs.dont_leave_empty_group, inits.with_group, "JsonLogic", (qb, onChange) => {
+  it("click on remove group will leave empty rule if canLeaveEmptyGroup=false", async () => {
+    await with_qb(configs.dont_leave_empty_group, inits.with_group, "JsonLogic", (qb, onChange) => {
       qb
         .find(".group--children .group .group--header .group--actions button")
         .at(2)
@@ -38,8 +38,8 @@ describe("interactions on vanilla", () => {
     });
   });
 
-  it("click on add rule will add new empty rule", () => {
-    with_qb(configs.simple_with_numbers_and_str, inits.with_number, "JsonLogic", (qb, onChange) => {
+  it("click on add rule will add new empty rule", async () => {
+    await with_qb(configs.simple_with_numbers_and_str, inits.with_number, "JsonLogic", (qb, onChange) => {
       qb
         .find(".group--actions button")
         .first()
@@ -50,8 +50,8 @@ describe("interactions on vanilla", () => {
     });
   });
 
-  it("click on add group will add new group with one empty rule if shouldCreateEmptyGroup=false", () => {
-    with_qb(configs.dont_leave_empty_group, inits.with_number, "JsonLogic", (qb, onChange) => {
+  it("click on add group will add new group with one empty rule if shouldCreateEmptyGroup=false", async () => {
+    await with_qb(configs.dont_leave_empty_group, inits.with_number, "JsonLogic", (qb, onChange) => {
       qb
         .find(".group--actions button")
         .at(1)
@@ -64,15 +64,16 @@ describe("interactions on vanilla", () => {
       expect(child.properties.conjunction).to.equal("AND"); //default
       const subchildKeys = Object.keys(child.children1);
       const subchild = child.children1[subchildKeys[0]];
-      expect(subchild).to.eql({
+      expect(JSON.stringify(subchild)).to.eql(JSON.stringify({
         type: "rule", 
-        properties: {field: null, operator: null, value: [], valueSrc: []}
-      });
+        id: subchild.id,
+        properties: {field: null, operator: null, value: [], valueSrc: []},
+      }));
     });
   });
 
-  it("change field to of same type will same op & value", () => {
-    with_qb(configs.simple_with_numbers_and_str, inits.with_number, "JsonLogic", (qb, onChange) => {
+  it("change field to of same type will same op & value", async () => {
+    await with_qb(configs.simple_with_numbers_and_str, inits.with_number, "JsonLogic", (qb, onChange) => {
       qb
         .find(".rule .rule--field select")
         .simulate("change", { target: { value: "num2" } });
@@ -86,8 +87,8 @@ describe("interactions on vanilla", () => {
     });
   });
 
-  it("change field to of another type will flush value and incompatible op", () => {
-    with_qb(configs.simple_with_numbers_and_str, inits.with_number, "JsonLogic", (qb, onChange) => {
+  it("change field to of another type will flush value and incompatible op", async () => {
+    await with_qb(configs.simple_with_numbers_and_str, inits.with_number, "JsonLogic", (qb, onChange) => {
       qb
         .find(".rule .rule--field select")
         .simulate("change", { target: { value: "str2" } });
@@ -101,8 +102,8 @@ describe("interactions on vanilla", () => {
     });
   });
 
-  it("change field to of another type will flush value and leave compatible op", () => {
-    with_qb(configs.simple_with_numbers_and_str, inits.with_number, "JsonLogic", (qb, onChange) => {
+  it("change field to of another type will flush value and leave compatible op", async () => {
+    await with_qb(configs.simple_with_numbers_and_str, inits.with_number, "JsonLogic", (qb, onChange) => {
       qb
         .find(".rule .rule--field select")
         .simulate("change", { target: { value: "str" } });
@@ -116,8 +117,8 @@ describe("interactions on vanilla", () => {
     });
   });
 
-  it("change field from group_ext to simple", () => {
-    with_qb(configs.with_group_array_cars, inits.with_group_array_cars, "JsonLogic", (qb, onChange) => {
+  it("change field from group_ext to simple", async () => {
+    await with_qb(configs.with_group_array_cars, inits.with_group_array_cars, "JsonLogic", (qb, onChange) => {
       qb
         .find(".rule_group_ext .group--field--count--rule .rule--field select")
         .simulate("change", { target: { value: "str" } });
@@ -131,8 +132,8 @@ describe("interactions on vanilla", () => {
     });
   });
 
-  it("change field from group_ext to simple with custom operator", () => {
-    with_qb(configs.with_group_array_custom_operator, inits.with_group_array_custom_operator, "JsonLogic", (qb, onChange) => {
+  it("change field from group_ext to simple with custom operator", async () => {
+    await with_qb(configs.with_group_array_custom_operator, inits.with_group_array_custom_operator, "JsonLogic", (qb, onChange) => {
       qb
         .find(".rule_group_ext .group--field--count--rule .rule--field select")
         .simulate("change", { target: { value: "str" } });
@@ -146,8 +147,8 @@ describe("interactions on vanilla", () => {
     });
   });
 
-  it("change field from simple to group_ext", () => {
-    with_qb(configs.with_group_array_cars, inits.with_text, "JsonLogic", (qb, onChange) => {
+  it("change field from simple to group_ext", async () => {
+    await with_qb(configs.with_group_array_cars, inits.with_text, "JsonLogic", (qb, onChange) => {
       qb
         .find(".rule .rule--field select")
         .simulate("change", { target: { value: "cars" } });
@@ -162,8 +163,8 @@ describe("interactions on vanilla", () => {
     });
   });
 
-  it("set not", () => {
-    with_qb(configs.simple_with_numbers_and_str, inits.with_number, "JsonLogic", (qb, onChange, {expect_jlogic}) => {
+  it("set not", async () => {
+    await with_qb(configs.simple_with_numbers_and_str, inits.with_number, "JsonLogic", (qb, onChange, {expect_jlogic}) => {
       qb
         .find('.group--conjunctions input[type="checkbox"]')
         .simulate("change", { target: { checked: true } });
@@ -173,8 +174,8 @@ describe("interactions on vanilla", () => {
     });
   });
 
-  it("change conjunction from AND to OR", () => {
-    with_qb(configs.simple_with_numbers_and_str, inits.with_2_numbers, "JsonLogic", (qb, onChange, {expect_jlogic}) => {
+  it("change conjunction from AND to OR", async () => {
+    await with_qb(configs.simple_with_numbers_and_str, inits.with_2_numbers, "JsonLogic", (qb, onChange, {expect_jlogic}) => {
       qb
         .find('.group--conjunctions input[type="radio"][value="OR"]')
         .simulate("change", { target: { value: "OR" } });
@@ -187,8 +188,8 @@ describe("interactions on vanilla", () => {
     });
   });
 
-  it("change value source to another field of same type", () => {
-    with_qb(configs.simple_with_numbers_and_str, inits.with_number, "JsonLogic", (qb, onChange, {expect_jlogic}) => {
+  it("change value source to another field of same type", async () => {
+    await with_qb(configs.simple_with_numbers_and_str, inits.with_number, "JsonLogic", (qb, onChange, {expect_jlogic}) => {
       qb
         .find(".rule .rule--value .widget--valuesrc select")
         .simulate("change", { target: { value: "field" } });
@@ -201,8 +202,8 @@ describe("interactions on vanilla", () => {
     });
   });
 
-  it("change op from equal to not_equal", () => {
-    with_qb(configs.simple_with_numbers_and_str, inits.with_number, "JsonLogic", (qb, onChange, {expect_jlogic}) => {
+  it("change op from equal to not_equal", async () => {
+    await with_qb(configs.simple_with_numbers_and_str, inits.with_number, "JsonLogic", (qb, onChange, {expect_jlogic}) => {
       qb
         .find(".rule .rule--operator select")
         .simulate("change", { target: { value: "not_equal" } });
@@ -212,8 +213,8 @@ describe("interactions on vanilla", () => {
     });
   });
 
-  it("remove group with 2 rules with confirm", () => {
-    with_qb(configs.with_settings_confirm, inits.with_number_and_group, "JsonLogic", (qb, onChange,  {expect_jlogic, config}) => {
+  it("remove group with 2 rules with confirm", async () => {
+    await with_qb(configs.with_settings_confirm, inits.with_number_and_group, "JsonLogic", (qb, onChange,  {expect_jlogic, config}) => {
       const renderConfirm = config.settings.renderConfirm;
       qb
         .find(".group--children .group .group--header .group--actions button")
@@ -224,8 +225,8 @@ describe("interactions on vanilla", () => {
     });
   });
 
-  it("remove group with 1 rule with confirm", () => {
-    with_qb(configs.with_settings_confirm, inits.with_number_and_group_1, "JsonLogic", (qb, onChange,  {expect_jlogic, config}) => {
+  it("remove group with 1 rule with confirm", async () => {
+    await with_qb(configs.with_settings_confirm, inits.with_number_and_group_1, "JsonLogic", (qb, onChange,  {expect_jlogic, config}) => {
       const renderConfirm = config.settings.renderConfirm;
       qb
         .find(".group--children .group .group--header .group--actions button")
@@ -236,8 +237,8 @@ describe("interactions on vanilla", () => {
     });
   });
 
-  it("remove rule with confirm", () => {
-    with_qb(configs.with_settings_confirm, inits.with_2_numbers, "JsonLogic", (qb, onChange,  {expect_jlogic, config}) => {
+  it("remove rule with confirm", async () => {
+    await with_qb(configs.with_settings_confirm, inits.with_2_numbers, "JsonLogic", (qb, onChange,  {expect_jlogic, config}) => {
       const renderConfirm = config.settings.renderConfirm;
       qb
         .find(".group--children .rule-container")
@@ -250,28 +251,28 @@ describe("interactions on vanilla", () => {
     });
   });
 
-  it("should not render not with showNot=false", () => {
-    with_qb(configs.with_settings_not_show_not, inits.with_numbers_and_group, "JsonLogic", (qb) => {
+  it("should not render not with showNot=false", async () => {
+    await with_qb(configs.with_settings_not_show_not, inits.with_numbers_and_group, "JsonLogic", (qb) => {
       expect(qb.find(".group--conjunctions input[type='checkbox']")).to.have.length(0);
     });
   });
 
-  it("should handle maxNumberOfRules=3", () => {
-    with_qb(configs.with_settings_max_number_of_rules_3, inits.with_number_and_group, "JsonLogic", (qb) => {
+  it("should handle maxNumberOfRules=3", async () => {
+    await with_qb(configs.with_settings_max_number_of_rules_3, inits.with_number_and_group, "JsonLogic", (qb) => {
       // if max exceeded --> (3) Add group | Add group, Delete group
       // else -->            (5) Add rule, Add group | Add rule, Add group, Delete group
       expect(qb.find(".group--actions button")).to.have.length(3);
     });
   });
 
-  it("should render labels with showLabels=true", () => {
-    with_qb([configs.with_different_groups, configs.with_settings_show_labels], inits.with_different_groups, "JsonLogic", (qb) => {
+  it("should render labels with showLabels=true", async () => {
+    await with_qb([configs.with_different_groups, configs.with_settings_show_labels], inits.with_different_groups, "JsonLogic", (qb) => {
       //todo
     });
   });
 
-  it("should render admin mode with showLock=true", () => {
-    with_qb([configs.with_different_groups, configs.with_settings_show_lock], inits.with_different_groups, "JsonLogic", (qb) => {
+  it("should render admin mode with showLock=true", async () => {
+    await with_qb([configs.with_different_groups, configs.with_settings_show_lock], inits.with_different_groups, "JsonLogic", (qb) => {
       //todo
     });
   });